+++ lib/freydis.rb
@@ -1,4 +1,8 @@
 require_relative 'freydis/version'
+require_relative 'freydis/options'
+require_relative 'freydis/init'
+require_relative 'freydis/data'
+
 module Freydis
 end