+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+* YAML config file in ~/.config/freydis/freydis.yaml
 
 ## 0.0.1, release 05/04/21
 * Initial push, code freeying !